# web/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Books API",
        "description": "Пример простого REST API",
        "contact": {
            "email": "iagmail@mail.ru"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Swagger Yii2 API Example (built in) - пример простого API"
        }
    ],
    "paths": {
...
        "/api/v1/book": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Список книг",
                "description": "Получение списка книг",
                "operationId": "listBooks",
                "parameters": [
                    {
                        "name": "expand",
                        "in": "query",
                        "description": "Дополнительные данные: createdAt, updatedAt"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Ограничение количества выводимых записей"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "''Сдвиг'' выборки от начала"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список книг",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Book"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Токен отсутствует или структура токена некорректна",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exception"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Время жизни токена истекло",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exception"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTAuthentification": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Добавление новой книги",
                "description": "Добавление новой книги",
                "operationId": "addBook",
                "requestBody": {
                    "description": "Добавление новой книги",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "author",
                                    "year"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Наименование"
                                    },
                                    "author": {
                                        "description": "Автор(ы)"
                                    },
                                    "year": {
                                        "description": "Год издания"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Книга добавлена"
                    },
                    "400": {
                        "description": "Токен отсутствует или структура токена некорректна",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exception"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Время жизни токена истекло",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exception"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибки валидации данных",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTAuthentification": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Exception": {
                "title": "Exception",
                "description": "Данные при возникновении исключения",
                "properties": {
                    "name": {
                        "description": "Наименование исключения",
                        "type": "string"
                    },
                    "message": {
                        "description": "Текст ошибки",
                        "type": "string"
                    },
                    "code": {
                        "description": "Код",
                        "type": "integer",
                        "enum": [
                            10,
                            20,
                            30,
                            40,
                            50,
                            60,
                            70
                        ]
                    },
                    "status": {
                        "description": "HTTP статус",
                        "type": "integer"
                    },
                    "type": {
                        "description": "Тип исключения",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Error": {
                "properties": {
                    "field": {
                        "description": "Поле",
                        "type": "string"
                    },
                    "message": {
                        "description": "Текст ошибки",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Book": {
                "title": "Book model",
                "description": "Book model",
                "required": [
                    "title",
                    "author",
                    "year"
                ],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "format": "int64"
                    },
                    "title": {
                        "title": "title",
                        "description": "Наименование",
                        "format": "string"
                    },
                    "author": {
                        "title": "author",
                        "description": "Автор(ы)",
                        "format": "string"
                    },
                    "year": {
                        "title": "year",
                        "description": "Год издания",
                        "format": "int32"
                    },
                    "createdAt": {
                        "title": "Создано",
                        "description": "Timestamp внесения данных",
                        "format": "datetime"
                    },
                    "updatedAt": {
                        "title": "Изменено",
                        "description": "Timestamp изменения данных",
                        "format": "datetime"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "JWTAuthentification": {
                "type": "http",
                "description": "Authentification by JWT token",
                "name": "JWT",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}